feat:YAML提交到 git 后，在stream上即可看到流水线 #6840
