+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/resources/user/UserStreamGitResourceImpl.kt
@@ -30,15 +30,8 @@ package com.tencent.devops.stream.resources.user
 import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.auth.api.AuthPermission
 import com.tencent.devops.common.client.Client
-import com.tencent.devops.common.pipeline.Model
-import com.tencent.devops.common.pipeline.container.Stage
-import com.tencent.devops.common.pipeline.container.TriggerContainer
-import com.tencent.devops.common.pipeline.pojo.element.trigger.ManualTriggerElement
 import com.tencent.devops.common.redis.RedisOperation
 import com.tencent.devops.common.web.RestResource
-import com.tencent.devops.process.engine.common.VMUtils
-import com.tencent.devops.process.pojo.setting.PipelineModelAndSetting
-import com.tencent.devops.process.pojo.setting.PipelineSetting
 import com.tencent.devops.project.api.service.ServiceProjectResource
 import com.tencent.devops.stream.api.user.UserStreamGitResource
 import com.tencent.devops.stream.constant.StreamConstant.STREAM_CI_FILE_DIR
@@ -56,10 +49,7 @@ import com.tencent.devops.stream.service.StreamGitTransferService
 import com.tencent.devops.stream.service.StreamPipelineService
 import com.tencent.devops.stream.service.StreamProjectService
 import com.tencent.devops.stream.service.StreamRequestService
-import com.tencent.devops.stream.trigger.actions.data.StreamTriggerPipeline
-import com.tencent.devops.stream.trigger.pojo.StreamTriggerLock
 import com.tencent.devops.stream.util.GitCommonUtils
-import com.tencent.devops.stream.util.StreamPipelineUtils
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 
@@ -159,86 +149,12 @@ class UserStreamGitResourceImpl @Autowired constructor(
             streamCreateFile = newFile
         )
         if (createNewFileIsSuccess) {
-            createNewPipeLine(gitProjectId, newFile, userId)
+             streamPipelineService.createNewPipeLine(gitProjectId, newFile.filePath, userId)
         }
         return Result(
             createNewFileIsSuccess
         )
     }
-
-    private fun createNewPipeLine(gitProjectId: String, newFile: StreamCreateFileInfo, userId: String) {
-        val pipeline = StreamTriggerPipeline(
-            gitProjectId = gitProjectId,
-            pipelineId = "",
-            filePath = newFile.filePath,
-            displayName = newFile.filePath,
-            enabled = true,
-            creator = userId
-        )
-        // pipelineId可能为blank所以使用filePath为key
-        val triggerLock = StreamTriggerLock(
-            redisOperation = redisOperation,
-            gitProjectId = gitProjectId,
-            filePath = pipeline.filePath
-        )
-        val gitProjectCode = getProjectCode(gitProjectId)
-        val realPipeline: StreamTriggerPipeline
-        // 避免出现多个触发拿到空的pipelineId后依次进来创建，所以需要在锁后重新获取pipeline
-        triggerLock.use {
-            triggerLock.lock()
-            realPipeline = getRealPipeLine(gitProjectId, pipeline)
-            // 优先创建流水线为了前台显示
-            if (realPipeline.pipelineId.isBlank()) {
-                streamPipelineService.savePipeline(
-                    pipeline = realPipeline,
-                    userId = userId,
-                    gitProjectId = gitProjectId.toLong(),
-                    projectCode = gitProjectCode,
-                    modelAndSetting = createTriggerModel(gitProjectCode),
-                    updateLastModifyUser = true
-                )
-            }
-        }
-    }
-
-    private fun getRealPipeLine(gitProjectId: String, pipeline: StreamTriggerPipeline) =
-        streamPipelineService.getPipelineByFile(
-            gitProjectId = gitProjectId.toLong(),
-            filePath = pipeline.filePath
-        )?.let {
-            StreamTriggerPipeline(it)
-        } ?: pipeline
-
-    private fun getProjectCode(gitProjectId: String): String {
-        return "git_$gitProjectId"
-    }
-
-    private fun createTriggerModel(projectCode: String) = PipelineModelAndSetting(
-        model = Model(
-            name = StreamPipelineUtils.genBKPipelineName(projectCode),
-            desc = "",
-            stages = listOf(
-                Stage(
-                    id = VMUtils.genStageId(1),
-                    name = VMUtils.genStageId(1),
-                    containers = listOf(
-                        TriggerContainer(
-                            id = "0",
-                            name = "构建触发",
-                            elements = listOf(
-                                ManualTriggerElement(
-                                    name = "手动触发",
-                                    id = "T-1-1-1"
-                                )
-                            )
-                        )
-                    )
-                )
-            )
-        ),
-        setting = PipelineSetting()
-    )
-
     // 默认在.ci目录下，.yml后缀
     private fun getFilePath(filePath: String): String {
         var newPath = filePath
