+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/service/StreamPipelineService.kt
@@ -31,12 +31,17 @@ import com.tencent.devops.common.api.pojo.Page
 import com.tencent.devops.common.api.util.PageUtil
 import com.tencent.devops.common.client.Client
 import com.tencent.devops.common.pipeline.Model
+import com.tencent.devops.common.pipeline.container.Stage
+import com.tencent.devops.common.pipeline.container.TriggerContainer
 import com.tencent.devops.common.pipeline.enums.ChannelCode
+import com.tencent.devops.common.pipeline.pojo.element.trigger.ManualTriggerElement
 import com.tencent.devops.common.redis.RedisLock
 import com.tencent.devops.common.redis.RedisOperation
 import com.tencent.devops.model.stream.tables.records.TGitPipelineResourceRecord
 import com.tencent.devops.process.api.service.ServicePipelineResource
+import com.tencent.devops.process.engine.common.VMUtils
 import com.tencent.devops.process.pojo.setting.PipelineModelAndSetting
+import com.tencent.devops.process.pojo.setting.PipelineSetting
 import com.tencent.devops.stream.config.StreamGitConfig
 import com.tencent.devops.stream.constant.StreamConstant
 import com.tencent.devops.stream.dao.GitPipelineResourceDao
@@ -46,6 +51,8 @@ import com.tencent.devops.stream.pojo.AllPathPair
 import com.tencent.devops.stream.pojo.StreamGitPipelineDir
 import com.tencent.devops.stream.pojo.StreamGitProjectPipeline
 import com.tencent.devops.stream.trigger.actions.data.StreamTriggerPipeline
+import com.tencent.devops.stream.trigger.pojo.StreamTriggerLock
+import com.tencent.devops.stream.util.StreamPipelineUtils
 import org.jooq.DSLContext
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -313,6 +320,80 @@ class StreamPipelineService @Autowired constructor(
             filePath = filePath
         )
     }
+
+    fun createNewPipeLine(gitProjectId: String, filePath: String, userId: String) {
+        val pipeline = StreamTriggerPipeline(
+            gitProjectId = gitProjectId,
+            pipelineId = "",
+            filePath = filePath,
+            displayName = filePath,
+            enabled = true,
+            creator = userId
+        )
+        // pipelineId可能为blank所以使用filePath为key
+        val triggerLock = StreamTriggerLock(
+            redisOperation = redisOperation,
+            gitProjectId = gitProjectId,
+            filePath = pipeline.filePath
+        )
+        val gitProjectCode = getProjectCode(gitProjectId)
+        val realPipeline: StreamTriggerPipeline
+        // 避免出现多个触发拿到空的pipelineId后依次进来创建，所以需要在锁后重新获取pipeline
+        triggerLock.use {
+            triggerLock.lock()
+            realPipeline = getRealPipeLine(gitProjectId, pipeline)
+            // 优先创建流水线为了前台显示
+            if (realPipeline.pipelineId.isBlank()) {
+                // 在蓝盾那边
+                savePipeline(
+                    pipeline = realPipeline,
+                    userId = userId,
+                    gitProjectId = gitProjectId.toLong(),
+                    projectCode = gitProjectCode,
+                    modelAndSetting = createTriggerModel(gitProjectCode),
+                    updateLastModifyUser = true
+                )
+            }
+        }
+    }
+
+    private fun getRealPipeLine(gitProjectId: String, pipeline: StreamTriggerPipeline) =
+        getPipelineByFile(
+            gitProjectId = gitProjectId.toLong(),
+            filePath = pipeline.filePath
+        )?.let {
+            StreamTriggerPipeline(it)
+        } ?: pipeline
+
+    private fun getProjectCode(gitProjectId: String): String {
+        return "git_$gitProjectId"
+    }
+
+    private fun createTriggerModel(projectCode: String) = PipelineModelAndSetting(
+        model = Model(
+            name = StreamPipelineUtils.genBKPipelineName(projectCode),
+            desc = "",
+            stages = listOf(
+                Stage(
+                    id = VMUtils.genStageId(1),
+                    name = VMUtils.genStageId(1),
+                    containers = listOf(
+                        TriggerContainer(
+                            id = "0",
+                            name = "构建触发",
+                            elements = listOf(
+                                ManualTriggerElement(
+                                    name = "手动触发",
+                                    id = "T-1-1-1"
+                                )
+                            )
+                        )
+                    )
+                )
+            )
+        ),
+        setting = PipelineSetting()
+    )
 
     private fun getPipelineLastBuildBranch(
         gitProjectId: Long,
